+++ src/App.tsx
@@ -1,7 +1,5 @@
 /* eslint-disable jsx-a11y/control-has-associated-label */
 import React, { useContext, useRef, useState } from 'react';
-import * as postServise from './api/todos';
-import { Errors } from './types/Errors';
 import { ErrorNotification } from './components/ErrorNotification';
 import { TodoHeader } from './components/TodoHeader';
 import { TodoList } from './components/TodoList';
@@ -13,33 +11,11 @@ export const App: React.FC = () => {
   const inputRef = useRef<HTMLInputElement>(null);
   const {
     todos,
-    setTodos,
     error,
-    setError,
   } = useContext(TodosContext);
 
   const handleChangeInput = (event: React.ChangeEvent<HTMLInputElement>) => {
     setTitle(event.target.value);
-  };
-
-  const toggleTodo = async (todoId: number) => {
-    const completedTodo = todos.map(todo => (
-      todo.id === todoId
-        ? { ...todo, completed: !todo.completed }
-        : { ...todo }
-    ));
-    const currentTodo = completedTodo.find(complted => complted.id === todoId);
-
-    try {
-      await postServise.updateTodo({
-        todo: currentTodo,
-        todoId,
-      });
-    } catch {
-      setError(Errors.UNABLE_UPDATE);
-    }
-
-    setTodos(completedTodo);
   };
 
   return (
@@ -54,9 +30,7 @@ export const App: React.FC = () => {
           handleChangeInput={handleChangeInput}
         />
 
-        <TodoList
-          toggleTodo={toggleTodo}
-        />
+        <TodoList />
 
         {!!todos.length && (
           <TodoFooter />
